defconfig: increase default root filesystem size
When trying to import some mongodb binary for further testing, the
archive does not fit in the remaining memory space. Bump the available
space to 1G to allow experimenting with mongodb

The major downside of this increase is the sd card write duration: if
this much space is indeed needed, some adjustments will be needed later.