overlay: embed couchdb server
Bring the couchdb server, used as a document-based NoSQL database to
store game data. Properly defining a buildroot package for couchdb has
proven to be a pain, after a few hours of trials:
- CouchDB is mostly written in Erlang
- it tries to import its own version of rebar
- I still did not get which rebar version it is trying to use
- it does not respect the passed cross-compilation variables

So let's hack our way through it and just bring a ready-to-use docker
image onto the system. This still makes sure that the image is imported
offline (no docker pull at first boot). To make it work, we need:
- an archive containing the whole couchdb image, stored in /var/lib
- a small script to detect if the image has already been imported or not
  into the local registry
- an "import" service, run only on first boot, thanks to the previous
  script

Once first boot is passed, the image is imported, and a container is
properly started. We let docker manage the automatic startup of the
CouchDB container.

Signed-off-by: Alexis Lothoré <alexis.lothore@gmail.com>