package/couchdb: enforce no replica
When the backend connects and creates the game DB, the following error
is generated by couchdb:

  Nov 09 14:46:42 nbc couchdb[403]: couchdb@127.0.0.1 <0.11812.0> 19a4e2c21c Request to create N=3 DB but only 1 node(s)
  Nov 09 14:46:42 nbc neon-beat-back[486]: 2025-11-09T14:46:42.457284Z  INFO neon_beat_back::services::storage_supervisor: storage reconnection succeeded after health check failure

This N=3 is the default number of replicas, that can be overriden by the
DB create request. The backend does not enforce this value, so this 3
value is taken, by our current couchdb configuration has no
configuration for replicas.

Enforce n=1 in couchdb configuration to get rid of this splat on DB
creation.

Signed-off-by: Alexis Lothoré <alexis.lothore@gmail.com>