kernel: bring patch fixing stability issue with Archer T3U usb dongle
When downloading a lot of data through the AP (eg: by downloading the
admin frontend page to a smartphone connected to the buzzers AP), the
connection eventually freezes. The kernel logs hint that something goes
wrong in the USB WiFi dongle:

  rtw88_8723du 1-2:1.2: error beacon valid
  rtw88_8723du 1-2:1.2: failed to download drv rsvd page
  rtw88_8723du 1-2:1.2: error beacon valid
  rtw88_8723du 1-2:1.2: failed to download drv rsvd page
  rtw88_8723du 1-2:1.2: failed to download beacon
  rtw88_8723du 1-2:1.2: error beacon valid

Fortunately, this issue has been identified upstream and fixed.
Unfortunately, it has not been backported in stable kernels. Bring in
the dedicated path to fix our kernel.

Signed-off-by: Alexis Lothoré <alexis.lothore@gmail.com>