scripts: update the components bump tool
The helper allowing to automatically bump component versions is able to
update both package version and SHA1, but it currently fails to compute
the correct SHA1 for the backend. The issue is due to the fact that when
buildroot downloads a cargo based project, it will "vendor" the
dependencies into the downloaded project (ie: it will run cargo vendor
--locked VENDOR in the project directory), and then retar the project
and compute its hash. Thanks to this, the dependencies are part of the
hash.

Follow the same strategy in the bump script. This is not operational
yet: the computed hash is still not the exact same one as computed by
buildroot, however we are getting closer: some manual inspection allowed
to validate that the directory content, after manual vendoring, is
exactly _the same_ as the buildroot one. So the failing step is either
in:
- the tar command
- the gzip command
- the hash computation

Signed-off-by: Alexis Lothoré <alexis.lothore@gmail.com>